+++ packages/common/src/util/http.js
@@ -2,6 +2,7 @@ import { Client, MockAgent } from 'undici';
 import { getReasonPhrase } from 'http-status-codes';
 import { Readable } from 'node:stream';
 import querystring from 'node:querystring';
+import path from 'node:path';
 
 const clients = new Map();
 
@@ -77,31 +78,57 @@ const assertOK = async (response, errorMap, fullUrl, method, startTime) => {
   }
 };
 
+export const ERROR_ABSOLUTE_URL = 'Absolute URLs not suppored';
+
+// throws if a path is absolute
+export const assertRelativeUrl = path => {
+  if (/https?:\/\//.test(path)) {
+    const e = new Error('UNEXPECTED_ABSOLUTE_URL');
+    e.code = 'UNEXPECTED_ABSOLUTE_URL';
+    e.description =
+      'An absolute URL was provided (https://...) but only a path (/a/b/c) is supported';
+    e.url = path;
+    e.fix =
+      'Remove the protocol, domain and origin from the provided URL. Maybe you need to use the generic HTTP helper functions instead?';
+    throw e;
+  }
+};
+
+export const ERROR_URL_MISMATCH = 'Target origin does not match baseUrl origin';
+
 export const parseUrl = (pathOrUrl = '', baseUrl) => {
   let fullUrl;
 
   // We handle our own URL parsing rather than leaning on node:url
   // because we are non-strict about the baseURL (ie, we do not ignore the path)
   if (/https?:\/\//.test(pathOrUrl)) {
-    fullUrl = pathOrUrl;
-  } else if (baseUrl) {
-    // ensure the base url ends with a /
-    const base = baseUrl.endsWith('/') ? baseUrl : baseUrl + '/';
-    // ensure the path does not start with /
-    const path = pathOrUrl.startsWith('/') ? pathOrUrl.substring(1) : pathOrUrl;
+    fullUrl = new URL(pathOrUrl);
 
-    fullUrl = base + path;
+    // If the url is absolute, and there's a basePath, ensure they point to the same origin
+    if (baseUrl) {
+      const base = new URL(baseUrl);
+      if (fullUrl.origin !== base.origin) {
+        const e = new Error(ERROR_URL_MISMATCH);
+        e.code = 'BASE_URL_MISMATCH';
+        e.description = `A request was attempted to an absolute URL, but a different base URL was specified. This is a potential security violation.`;
+        e.target = pathOrUrl;
+        e.baseUrl = baseUrl;
+        e.fix = 'Try using a generic HTTP function to access the target URL';
+        throw e;
+      }
+    }
+  } else if (baseUrl) {
+    fullUrl = new URL(path.join(baseUrl, pathOrUrl));
   } else {
     // let this throw
     new URL(pathOrUrl);
   }
 
-  const url = new URL(fullUrl);
   return {
-    url: url.toString(),
-    baseUrl: url.origin,
-    path: url.pathname,
-    query: querystring.parse(url.searchParams.toString()),
+    url: fullUrl.toString(),
+    baseUrl: fullUrl.origin,
+    path: fullUrl.pathname,
+    query: querystring.parse(fullUrl.searchParams.toString()),
   };
 };