+++ packages/magpi/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-magpi
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [4fe527c]
+  - @openfn/language-common@2.0.0
+
 ## 1.2.0
 
 ### Minor Changes