+++ packages/mysql/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @openfn/language-mysql
 
+## 2.0.0
+
+### Major Changes
+
+- Export new common http helpers (http namespace)
+
+## 1.5.3
+
+### Patch Changes
+
+- Updated dependencies [4fe527c]
+  - @openfn/language-common@2.0.0
+
 ## 1.5.2
 
 ### Patch Changes