+++ packages/postgresql/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @openfn/language-postgresql
 
+## 6.0.0
+
+### Major Changes
+
+- Export new common http helpers (http namespace)
+
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [4fe527c]
+  - @openfn/language-common@2.0.0
+
 ## 5.0.1
 
 ### Patch Changes